refactor: 더미 관리 기능 package 구조 리팩토링 및 예외 처리 개선
- `fittoring.application.community.dummy` → `fittoring.admin` 패키지로 이동
  - DTO: `fittoring.admin.presentation.dto`
  - Service, Repository: `fittoring.admin.service`, `fittoring.admin.repository`
  - Custom Exceptions: `fittoring.admin.exception`
- `GlobalExceptionHandler`에 커스텀 예외 경로 수정 적용
- `scenarios1.yml`에서 `scheduled_at` 값 일괄 수정 (1시간씩 증가)
