+++ backend/fittoring/src/main/java/fittoring/admin/exception/DummyAlreadyInsertedException.java
@@ -1,4 +1,4 @@
-package fittoring.application.community.dummy;
+package fittoring.admin.exception;
 
 public class DummyAlreadyInsertedException extends RuntimeException {
