+++ backend/fittoring/src/main/java/fittoring/admin/exception/DummyScenarioFileNotFoundException.java
@@ -1,4 +1,4 @@
-package fittoring.application.community.dummy;
+package fittoring.admin.exception;
 
 public class DummyScenarioFileNotFoundException extends RuntimeException {
