+++ backend/fittoring/src/main/java/fittoring/admin/exception/InvalidDummyScenarioException.java
@@ -1,4 +1,4 @@
-package fittoring.application.community.dummy;
+package fittoring.admin.exception;
 
 public class InvalidDummyScenarioException extends RuntimeException {
