+++ backend/fittoring/src/main/java/fittoring/exception/GlobalExceptionHandler.java
@@ -43,9 +43,9 @@ import fittoring.application.exception.ReviewNotFoundException;
 import fittoring.application.exception.UnauthorizedChatMessageAccessException;
 import fittoring.application.exception.UnauthorizedChatRoomAccessException;
 import fittoring.application.exception.UnauthorizedException;
-import fittoring.application.community.dummy.DummyAlreadyInsertedException;
-import fittoring.application.community.dummy.DummyScenarioFileNotFoundException;
-import fittoring.application.community.dummy.InvalidDummyScenarioException;
+import fittoring.admin.exception.DummyAlreadyInsertedException;
+import fittoring.admin.exception.DummyScenarioFileNotFoundException;
+import fittoring.admin.exception.InvalidDummyScenarioException;
 import fittoring.application.exception.UnsupportedImageExtensionException;
 import fittoring.infrastructure.exception.S3UploadException;
 import fittoring.infrastructure.exception.SmsException;
